split file to seperate go file
