+++ codegen/data.go
@@ -1,0 +1,7 @@
+package main
+
+type Data struct {
+	Price    string
+	Currency string
+	Reuse    string
+}
