+++ codegen/payload.go
@@ -1,0 +1,7 @@
+package main
+
+type Payload struct {
+	Status string
+	Key    string
+	Value  Data
+}
